PHP Integration
Started the JS/PHP integration with the mockup. Had to change some
dependencies in the mockup, and changed some var names for the forms for
PHP.

Currently, if you go to login and your username is not in the database,
the PHP redirects you back to the home page. Will need to add an alert
box or something saying "Unsuccessful login or something".

The same applies to the register php file.

Split Ireland Init into separate php files.